Add client-common crate to avoid circular dependencies
