+++ client-common/Cargo.toml
@@ -1,38 +1,33 @@
 [package]
-name = "solana-rpc-client"
+name = "solana-client-common"
 version = "1.12.0"
-description = "Solana RPC Client"
+description = "Solana Client Common Utilities"
 authors = ["Solana Maintainers <maintainers@solana.foundation>"]
 repository = "https://github.com/solana-labs/solana"
 license = "Apache-2.0"
 homepage = "https://solana.com/"
-documentation = "https://docs.rs/solana-rpc-client"
+documentation = "https://docs.rs/solana-client-common"
 edition = "2021"
 
 [dependencies]
-async-trait = "0.1.57"
 base64 = "0.13.0"
-bincode = "1.3.3"
 bs58 = "0.4.0"
-log = "0.4.17"
+indicatif = "0.17.0"
+jsonrpc-core = "18.0.0"
 reqwest = { version = "0.11.11", default-features = false, features = ["blocking", "brotli", "deflate", "gzip", "rustls-tls", "json"] }
 semver = "1.0.13"
 serde = "1.0.143"
 serde_derive = "1.0.103"
 serde_json = "1.0.83"
 solana-account-decoder = { path = "../account-decoder", version = "=1.12.0" }
-solana-client-common = { path = "../client-common", version = "=1.12.0" }
+solana-faucet = { path = "../faucet", version = "=1.12.0" }
 solana-sdk = { path = "../sdk", version = "=1.12.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.12.0" }
 solana-version = { path = "../version", version = "=1.12.0" }
-solana-vote-program = { path = "../programs/vote", version = "=1.12.0" }
-tokio = { version = "1", features = ["full"] }
+spl-token-2022 = { version = "=0.4.3", features = ["no-entrypoint"] }
+thiserror = "1.0"
 
 [dev-dependencies]
-assert_matches = "1.5.0"
-crossbeam-channel = "0.5"
-jsonrpc-core = "18.0.0"
-jsonrpc-http-server = "18.0.0"
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
