+++ client-common/src/rpc_error_object.rs
