+++ client/Cargo.toml
@@ -41,6 +41,7 @@ serde_derive = "1.0.103"
 serde_json = "1.0.83"
 solana-account-decoder = { path = "../account-decoder", version = "=1.12.0" }
 solana-clap-utils = { path = "../clap-utils", version = "=1.12.0" }
+solana-client-common = { path = "../client-common", version = "=1.12.0" }
 solana-faucet = { path = "../faucet", version = "=1.12.0" }
 solana-measure = { path = "../measure", version = "=1.12.0" }
 solana-metrics = { path = "../metrics", version = "=1.12.0" }
