+++ pubsub-client/Cargo.toml
@@ -19,7 +19,7 @@ serde = "1.0.143"
 serde_derive = "1.0.103"
 serde_json = "1.0.83"
 solana-account-decoder = { path = "../account-decoder", version = "=1.12.0" }
-solana-client = { path = "../client", version = "=1.12.0" }
+solana-client-common = { path = "../client-common", version = "=1.12.0" }
 solana-sdk = { path = "../sdk", version = "=1.12.0" }
 thiserror = "1.0"
 tokio = { version = "1", features = ["full"] }
