+++ thin-client/Cargo.toml
@@ -12,7 +12,7 @@ edition = "2021"
 [dependencies]
 bincode = "1.3.3"
 log = "0.4.17"
-solana-client = { path = "../client", version = "=1.12.0" }
+solana-client-common = { path = "../client-common", version = "=1.12.0" }
 solana-rpc-client = { path = "../rpc-client", version = "=1.12.0" }
 solana-sdk = { path = "../sdk", version = "=1.12.0" }
 solana-tpu-client = { path = "../tpu-client", version = "=1.12.0" }
