+++ thin-client/src/thin_client.rs
@@ -5,7 +5,7 @@
 
 use {
     log::*,
-    solana_client::{rpc_config::RpcProgramAccountsConfig, rpc_response::Response},
+    solana_client_common::{rpc_config::RpcProgramAccountsConfig, rpc_response::Response},
     solana_rpc_client::rpc_client::RpcClient,
     solana_sdk::{
         account::Account,
