+++ tpu-client/Cargo.toml
@@ -26,7 +26,7 @@ quinn-udp = "0.1.3"
 rand = "0.7.0"
 rayon = "1.5.3"
 rustls = { version = "0.20.6", features = ["dangerous_configuration"] }
-solana-client = { path = "../client", version = "=1.12.0" }
+solana-client-common = { path = "../client-common", version = "=1.12.0" }
 solana-measure = { path = "../measure", version = "=1.12.0" }
 solana-metrics = { path = "../metrics", version = "=1.12.0" }
 solana-net-utils = { path = "../net-utils", version = "=1.12.0" }
